[Costing] Testing memory consumption of polymorphic builtins
