+++ plutus-core/src/Language/PlutusCore/Evaluation/Machine/ExMemory.hs
@@ -63,6 +63,7 @@ newtype ExMemory = ExMemory Integer
 -- Yes, I also hate that.
 deriving newtype instance PrettyDefaultBy config Integer => PrettyBy config ExMemory
 
+-- 'Integer's are not particularly fast. Should we use @Int64@?
 -- | Counts CPU units - no fixed base, proportional.
 newtype ExCPU = ExCPU Integer
   deriving (Eq, Ord, Show)
