+++ plutus-core/test/Evaluation/DynamicBuiltins/Definition.hs
@@ -1,20 +1,7 @@
 -- | A dynamic built-in name test.
 
-{-# OPTIONS_GHC -fno-warn-orphans               #-}
-{-# OPTIONS_GHC -fno-warn-redundant-constraints #-}
-
-{-# LANGUAGE AllowAmbiguousTypes   #-}
-{-# LANGUAGE DataKinds             #-}
-{-# LANGUAGE DeriveAnyClass        #-}
-{-# LANGUAGE DerivingVia           #-}
-{-# LANGUAGE FlexibleInstances     #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings     #-}
-{-# LANGUAGE PolyKinds             #-}
-{-# LANGUAGE TypeApplications      #-}
-{-# LANGUAGE TypeFamilies          #-}
-{-# LANGUAGE TypeOperators         #-}
-{-# LANGUAGE UndecidableInstances  #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeApplications  #-}
 
 module Evaluation.DynamicBuiltins.Definition
     ( test_definition
@@ -22,168 +9,23 @@ module Evaluation.DynamicBuiltins.Definition
 
 import           Language.PlutusCore
 import           Language.PlutusCore.Constant
-import           Language.PlutusCore.Evaluation.Machine.ExBudgeting
-import           Language.PlutusCore.Evaluation.Machine.ExBudgetingDefaults
-import           Language.PlutusCore.Evaluation.Machine.ExMemory
-import           Language.PlutusCore.Evaluation.Machine.Exception
 import           Language.PlutusCore.Generators.Interesting
-import           Language.PlutusCore.MkPlc                                  hiding (error)
-import           Language.PlutusCore.Pretty
+import           Language.PlutusCore.MkPlc                  hiding (error)
 
+import           Language.PlutusCore.Examples.Builtins
 import           Language.PlutusCore.StdLib.Data.Bool
-import qualified Language.PlutusCore.StdLib.Data.Function                   as Plc
-import qualified Language.PlutusCore.StdLib.Data.List                       as Plc
+import qualified Language.PlutusCore.StdLib.Data.Function   as Plc
+import qualified Language.PlutusCore.StdLib.Data.List       as Plc
 
 import           Evaluation.DynamicBuiltins.Common
 
 import           Data.Either
-import qualified Data.Kind                                                  as GHC (Type)
 import           Data.Proxy
-import           Data.Text.Prettyprint.Doc
-import           Data.Void
-import           GHC.Generics
-import           GHC.Ix
-import           Hedgehog                                                   hiding (Opaque, Size, Var)
-import qualified Hedgehog.Gen                                               as Gen
+import           Hedgehog                                   hiding (Opaque, Size, Var)
+import qualified Hedgehog.Gen                               as Gen
 import           Test.Tasty
 import           Test.Tasty.HUnit
 import           Test.Tasty.Hedgehog
-
-instance (Bounded a, Bounded b) => Bounded (Either a b) where
-    minBound = Left  minBound
-    maxBound = Right maxBound
-
-size :: forall a. (Bounded a, Enum a) => Int
-size = fromEnum (maxBound :: a) - fromEnum (minBound :: a) + 1
-
--- >>> map fromEnum [Left False .. Right GT]
--- [0,1,2,3,4]
--- >>> map toEnum [0 .. 4] :: [Either Bool Ordering]
--- [Left False,Left True,Right LT,Right EQ,Right GT]
-instance (Eq a, Eq b, Bounded a, Bounded b, Enum a, Enum b) => Enum (Either a b) where
-    succ (Left x)
-        | x == maxBound = Right minBound
-        | otherwise     = Left $ succ x
-    succ (Right y)
-        | y == maxBound = error "Out of bounds"
-        | otherwise     = Right $ succ y
-
-    pred (Left x)
-        | x == minBound = error "Out of bounds"
-        | otherwise     = Left $ pred x
-    pred (Right y)
-        | y == minBound = Left maxBound
-        | otherwise     = Right $ pred y
-
-    toEnum i
-        | i < s     = Left  $ toEnum i
-        | otherwise = Right $ toEnum (i - s)
-        where s = size @a
-
-    fromEnum (Left  x) = fromEnum x
-    fromEnum (Right y) = size @a + fromEnum y
-
--- >>> import GHC.Ix
--- >>> map (unsafeIndex (Left False, Right GT)) [Left False .. Right GT]
--- [0,1,2,3,4]
--- >>> let bounds = (Left (False, EQ), Right (True, GT)) in map (unsafeIndex bounds) $ range bounds
--- [0,1,2,3,4,5,6,7,8,9]
-instance (Bounded a, Bounded b, Ix a, Ix b) => Ix (Either a b) where
-    range (Right _, Left  _) = []
-    range (Right x, Right y) = map Right (range (x, y))
-    range (Left  x, Right y) = map Left (range (x, maxBound)) ++ map Right (range (minBound, y))
-    range (Left  x, Left  y) = map Left (range (x, y))
-
-    unsafeIndex (Right _, _) (Left  _) = error "Out of bounds"
-    unsafeIndex (Left  x, n) (Left  i) = unsafeIndex (x, fromLeft maxBound n) i
-    unsafeIndex (Right x, n) (Right i) = unsafeIndex (x, fromRight (error "Out of bounds") n) i
-    unsafeIndex (Left  x, n) (Right i) =
-        unsafeIndex (x, maxBound) maxBound + 1 +
-            unsafeIndex (minBound, fromRight (error "Out of bounds") n) i
-
-    inRange (m, n) i = m <= i && i <= n
-
-data ExtensionFun
-    = Factorial
-    | Const
-    | Id
-    | IdFInteger
-    | IdList
-    | IdRank2
-    | Absurd
-    deriving (Show, Eq, Ord, Enum, Bounded, Ix, Generic, Hashable)
-    deriving ExMemoryUsage via (GenericExMemoryUsage ExtensionFun)
-
-instance Pretty ExtensionFun where pretty = viaShow
-
-instance (ToBuiltinMeaning uni fun1, ToBuiltinMeaning uni fun2) =>
-            ToBuiltinMeaning uni (Either fun1 fun2) where
-    type DynamicPart uni (Either fun1 fun2) = (DynamicPart uni fun1, DynamicPart uni fun2)
-    type CostingPart uni (Either fun1 fun2) = (CostingPart uni fun1, CostingPart uni fun2)
-
-    toBuiltinMeaning (Left  fun) = case toBuiltinMeaning fun of
-        BuiltinMeaning sch toF toExF -> BuiltinMeaning sch (toF . fst) (toExF . fst)
-    toBuiltinMeaning (Right fun) = case toBuiltinMeaning fun of
-        BuiltinMeaning sch toF toExF -> BuiltinMeaning sch (toF . snd) (toExF . snd)
-
-defBuiltinsRuntimeExt
-    :: HasConstantIn DefaultUni term
-    => BuiltinsRuntime (Either DefaultFun ExtensionFun) term
-defBuiltinsRuntimeExt = toBuiltinsRuntime mempty (defaultCostModel, ())
-
-data ListRep (a :: GHC.Type)
-instance KnownTypeAst uni a => KnownTypeAst uni (ListRep a) where
-    toTypeAst _ = TyApp () Plc.listTy . toTypeAst $ Proxy @a
-type instance ToBinds (ListRep a) = ToBinds a
-
-instance KnownTypeAst uni Void where
-    toTypeAst _ = runQuote $ do
-        a <- freshTyName "a"
-        pure $ TyForall () a (Type ()) $ TyVar () a
-instance KnownType term Void where
-    makeKnown = absurd
-    readKnown = throwingWithCause _UnliftingError "Can't unlift a 'Void'" . Just
-type instance ToBinds Void = '[]
-
-instance (GShow uni, GEq uni, uni `Includes` Integer) => ToBuiltinMeaning uni ExtensionFun where
-    type DynamicPart uni ExtensionFun = ()
-    type CostingPart uni ExtensionFun = ()
-    toBuiltinMeaning Factorial = toStaticBuiltinMeaning (\(n :: Integer) -> product [1..n]) mempty
-    toBuiltinMeaning Const =
-        toStaticBuiltinMeaning
-            const
-            mempty
-    toBuiltinMeaning Id =
-        toStaticBuiltinMeaning
-            Prelude.id
-            mempty
-    toBuiltinMeaning IdFInteger =
-        toStaticBuiltinMeaning
-            (Prelude.id
-                :: a ~ Opaque term (TyAppRep (TyVarRep ('TyNameRep "f" 0)) Integer)
-                => a -> a)
-            mempty
-    toBuiltinMeaning IdList =
-        toStaticBuiltinMeaning
-            (Prelude.id
-                :: a ~ Opaque term (ListRep (TyVarRep ('TyNameRep "a" 0)))
-                => a -> a)
-            mempty
-    toBuiltinMeaning IdRank2 =
-        toStaticBuiltinMeaning
-            (Prelude.id
-                :: ( f ~ 'TyNameRep "f" 0
-                   , a ~ 'TyNameRep @GHC.Type "a" 1
-                   , afa ~ Opaque term (TyForallRep a (TyAppRep (TyVarRep f) (TyVarRep a)))
-                   )
-                => afa -> afa)
-            mempty
-    toBuiltinMeaning Absurd =
-        toStaticBuiltinMeaning
-            (absurd
-                :: a ~ Opaque term (TyVarRep ('TyNameRep "a" 0))
-                => Void -> a)
-            mempty
 
 -- | Check that 'Factorial' from the above computes to the same thing as
 -- a factorial defined in PLC itself.
@@ -240,11 +82,7 @@ test_Id =
 test_IdFInteger :: TestTree
 test_IdFInteger =
     testCase "IdFInteger" $ do
-        let tyAct = typeOfBuiltinFunction @DefaultUni IdFInteger
-            tyExp = let f = TyName . Name "f" $ Unique 0
-                        fInteger = TyApp () (TyVar () f) $ mkTyBuiltin @Integer ()
-                    in TyForall () f (KindArrow () (Type ()) $ Type ()) $ TyFun () fInteger fInteger
-            one = mkConstant @Integer @DefaultUni () 1
+        let one = mkConstant @Integer @DefaultUni () 1
             ten = mkConstant @Integer @DefaultUni () 10
             res = mkConstant @Integer @DefaultUni () 55
             -- sum (idFInteger {list} (enumFromTo 1 10))
@@ -252,7 +90,6 @@ test_IdFInteger =
                 = Apply () (mapFun Left Plc.sum)
                 . Apply () (TyInst () (Builtin () $ Right IdFInteger) Plc.listTy)
                 $ mkIterApp () (mapFun Left Plc.enumFromTo) [one, ten]
-        tyAct @?= tyExp
         typecheckEvaluateCkNoEmit defBuiltinsRuntimeExt term @?= Right (EvaluationSuccess res)
 
 test_IdList :: TestTree
@@ -302,30 +139,14 @@ argument when it's a function, for another example).
 test_IdRank2 :: TestTree
 test_IdRank2 =
     testCase "IdRank2" $ do
-        let tyAct = typeOfBuiltinFunction @DefaultUni IdRank2
-            tyExp = let f = TyName . Name "f" $ Unique 0
-                        a = TyName . Name "a" $ Unique 1
-                        allAfA = TyForall () a (Type ()) . TyApp () (TyVar () f) $ TyVar () a
-                    in TyForall () f (KindArrow () (Type ()) $ Type ()) $ TyFun () allAfA allAfA
-            res = mkConstant @Integer @DefaultUni () 0
+        let res = mkConstant @Integer @DefaultUni () 0
             integer = mkTyBuiltin @Integer ()
             -- sum (idRank2 {list} nil {integer})
             term
                 = Apply () (mapFun Left Plc.sum)
                 . TyInst () (Apply () (TyInst () (Builtin () $ Right IdRank2) Plc.listTy) Plc.nil)
                 $ integer
-        tyAct @?= tyExp
         typecheckEvaluateCkNoEmit defBuiltinsRuntimeExt term @?= Right (EvaluationSuccess res)
-
--- | Test that the type of PLC @Absurd@ is inferred correctly.
-test_Absurd :: TestTree
-test_Absurd =
-    testCase "Absurd" $ do
-        let tyAct = typeOfBuiltinFunction @DefaultUni Absurd
-            tyExp = let a = TyName . Name "a" $ Unique 0
-                        tyForallA = TyForall () a (Type ())
-                    in tyForallA . TyFun () (tyForallA $ TyVar () a) $ TyVar () a
-        tyAct @?= tyExp
 
 test_definition :: TestTree
 test_definition =
@@ -336,5 +157,4 @@ test_definition =
         , test_IdFInteger
         , test_IdList
         , test_IdRank2
-        , test_Absurd
         ]
