+++ plutus-core/test/TypeSynthesis/Golden/Absurd.plc.golden
@@ -1,0 +1,1 @@
+(all a (type) (fun (all a (type) a) a))
