+++ plutus-core/test/TypeSynthesis/Golden/Const.plc.golden
@@ -1,0 +1,1 @@
+(all a (type) (all b (type) (fun a (fun b a))))
