+++ plutus-core/test/TypeSynthesis/Golden/Factorial.plc.golden
@@ -1,0 +1,1 @@
+(fun (con integer) (con integer))
