+++ plutus-core/test/TypeSynthesis/Golden/Id.plc.golden
@@ -1,0 +1,1 @@
+(all a (type) (fun a a))
