+++ plutus-core/test/TypeSynthesis/Golden/IdFInteger.plc.golden
@@ -1,0 +1,1 @@
+(all f (fun (type) (type)) (fun [f (con integer)] [f (con integer)]))
