+++ plutus-core/test/TypeSynthesis/Golden/IdList.plc.golden
@@ -1,0 +1,1 @@
+(all a (type) (fun [(lam a (type) (ifix (lam list (fun (type) (type)) (lam a (type) (all r (type) (fun r (fun (fun a (fun [list a] r)) r))))) a)) a] [(lam a (type) (ifix (lam list (fun (type) (type)) (lam a (type) (all r (type) (fun r (fun (fun a (fun [list a] r)) r))))) a)) a]))
