+++ plutus-core/test/TypeSynthesis/Golden/IdRank2.plc.golden
@@ -1,0 +1,1 @@
+(all f (fun (type) (type)) (fun (all a (type) [f a]) (all a (type) [f a])))
