+++ plutus-core/untyped-plutus-core-test/Evaluation/Machines/Budget/IdNat/0.plc.golden
@@ -1,0 +1,23 @@
+( (Right (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> z)))))))))))))))))))))))
+, ({ tally: ({ BVar causes ({ cpu: 140
+| mem: 0
+})
+| BDelay causes ({ cpu: 37
+| mem: 0
+})
+| BLamAbs causes ({ cpu: 94
+| mem: 0
+})
+| BForce causes ({ cpu: 26
+| mem: 0
+})
+| BApply causes ({ cpu: 122
+| mem: 0
+})
+| BAST causes ({ cpu: 0
+| mem: 88
+})})
+| budget: ({ cpu: 419
+| mem: 88
+})
+}) )
