+++ plutus-core/untyped-plutus-core-test/Evaluation/Machines/Budget/IdNat/3.plc.golden
@@ -1,0 +1,29 @@
+( (Right (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> f (delay (\z -> \f -> z)))))))))))))))))))))))
+, ({ tally: ({ BBuiltin causes ({ cpu: 3
+| mem: 0
+})
+| BVar causes ({ cpu: 140
+| mem: 0
+})
+| BDelay causes ({ cpu: 37
+| mem: 0
+})
+| BLamAbs causes ({ cpu: 94
+| mem: 0
+})
+| BBuiltinApp Id causes ({ cpu: 3
+| mem: 0
+})
+| BForce causes ({ cpu: 29
+| mem: 0
+})
+| BApply causes ({ cpu: 125
+| mem: 0
+})
+| BAST causes ({ cpu: 0
+| mem: 91
+})})
+| budget: ({ cpu: 431
+| mem: 91
+})
+}) )
